extract mutators into useTicketPriceCalculatorFormMutators
